Refactor file to relevant component folders
