+++ app/components/main.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import App from './app.jsx';
+import Nav from './nav-bar/app.jsx';
 
 export default class Main extends React.Component {
   constructor() {
@@ -9,7 +9,7 @@ export default class Main extends React.Component {
   render() {
     return (
       <div>
-        <App />
+        <Nav />
         <div className="">
           {this.props.children}
         </div>
