--- app/components/nav.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default class LeftNav extends React.Component {
-
-}
