+++ app/main.js
@@ -2,8 +2,8 @@ import React from 'react';
 import {Router, Route, Link, browserHistory, IndexRoute} from 'react-router';
 import {render} from 'react-dom';
 import App from './components/main.jsx';
-import Documents from './components/document-list.jsx';
-// import Docs from './components/doc-list.jsx';
+import Documents from './components/landing-page/document-list.jsx';
+
 require('./styles/style.css');
 
 render((
