+++ app/npm-debug.log
@@ -1,0 +1,42 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'start' ]
+2 info using npm@3.3.6
+3 info using node@v5.0.0
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle doc-management-system@1.0.0~prestart: doc-management-system@1.0.0
+6 silly lifecycle doc-management-system@1.0.0~prestart: no script for prestart, continuing
+7 info lifecycle doc-management-system@1.0.0~start: doc-management-system@1.0.0
+8 verbose lifecycle doc-management-system@1.0.0~start: unsafe-perm in lifecycle true
+9 verbose lifecycle doc-management-system@1.0.0~start: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/bryomckim/Andela/DMS-Frontend/node_modules/.bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin
+10 verbose lifecycle doc-management-system@1.0.0~start: CWD: /Users/bryomckim/Andela/DMS-Frontend
+11 silly lifecycle doc-management-system@1.0.0~start: Args: [ '-c', 'node server.js' ]
+12 silly lifecycle doc-management-system@1.0.0~start: Returned: code: 1  signal: null
+13 info lifecycle doc-management-system@1.0.0~start: Failed to exec start script
+14 verbose stack Error: doc-management-system@1.0.0 start: `node server.js`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:233:16)
+14 verbose stack     at emitTwo (events.js:87:13)
+14 verbose stack     at EventEmitter.emit (events.js:172:7)
+14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:24:14)
+14 verbose stack     at emitTwo (events.js:87:13)
+14 verbose stack     at ChildProcess.emit (events.js:172:7)
+14 verbose stack     at maybeClose (internal/child_process.js:818:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:211:5)
+15 verbose pkgid doc-management-system@1.0.0
+16 verbose cwd /Users/bryomckim/Andela/DMS-Frontend/app
+17 error Darwin 15.3.0
+18 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "start"
+19 error node v5.0.0
+20 error npm  v3.3.6
+21 error code ELIFECYCLE
+22 error doc-management-system@1.0.0 start: `node server.js`
+22 error Exit status 1
+23 error Failed at the doc-management-system@1.0.0 start script 'node server.js'.
+23 error This is most likely a problem with the doc-management-system package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     node server.js
+23 error You can get their info via:
+23 error     npm owner ls doc-management-system
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
